net: netcp: ethss: Do not read 10gbe pcsr reg if no serdes phy defined
If there is no serdes phy defined, i.e. no individual SerDes
lane enabling is allowed, there is no point to read the PCSR
status register to determine the link status. This is the case
when 10gbe MCU firmware is used in configuring the 10gbe
SerDes. This patch implements the above mentioned changes.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Michael Scherban <m-scherban@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>