hsr: update to support lre rx and l2 forward offload
hsr LRE functionality as well L2 forwarding is expected to be offloaded
to hardware (h/w) or firmware (f/w). When the LRE Ingress function is
offloaded, the LRE function of duplicating and tagging in egress direction
is still handled in the hsr driver.  At the ingress, the supervisor
frames are processed at h/w or f/w and is responsible for maintaining the
node table for remote nodes and statistics. h/w or f/w is also expected to
drop duplicate of the received frames to host and forward only one copy to
slave A or Slave B depending on which interface frame came first. Non Host
and broadcast/multicast frames gets forwarded through hsr L2 forwarder
code if NETIF_F_HW_L2FW_DOFFLOAD is not supported. If supported, then no
forwarding happens in the hsr L2 forwarder and the same is assumed to be
handled at h/w or f/w.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>